Move Status debug formatter into testing too
