+++ stellar_contract_sdk/src/val.rs
@@ -160,7 +160,7 @@ impl Val {
     }
 
     #[inline(always)]
-    pub(crate) fn get_body(&self) -> u64 {
+    pub(crate) const fn get_body(&self) -> u64 {
         self.0 >> 4
     }
 
@@ -300,18 +300,5 @@ impl Val {
     #[inline(always)]
     pub fn from_object(obj: Object) -> Val {
         obj.into()
-    }
-}
-
-#[cfg(not(target_family = "wasm"))]
-use core::fmt::Debug;
-
-#[cfg(not(target_family = "wasm"))]
-impl Debug for Val {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("Val")
-            .field("body", &self.get_body())
-            .field("tag", &self.get_tag())
-            .finish()
     }
 }
